Add and use public character ID as resource ID over character name
This removes the need to validate and normalize character names, as
anything is now acceptable.

An idempotency key has also been introduced to avoid duplicate character
creations. This is enforce through the use of the idempotency key as the
character's public ID, thus creating a uniqueness constraint violation
on duplicate entries. Note that the creation endpoint currently returns
a `409 Conflict` response on duplicate entries. An alternative would be
to pretend as if the resource was just created and return a
`201 Created` response instead, but this would be misleading if the
request is not actually identical (even though the idempotency key is).