Improve idempotent handling
Return `201 Created` on duplicate character creation when request is
"identical" (only the parameters relevant to the command are inspected).

The implementation is optimistic (it assumes that a uniqueness
constraint violation rarely occurs) and tries to write to the database
without checking existing entries. This off-loads the extra database
round-trip required to "properly" handle idempotency to the rare
situations where a uniqueness constraint violation does occur.